+++ packages/wrangler/e2e/seed-files/mixed-mode-workers/simple-service-binding.js
@@ -1,0 +1,8 @@
+export default {
+	async fetch(request, env) {
+		const remoteWorkerText = await (
+			await env.REMOTE_WORKER.fetch(request)
+		).text();
+		return new Response(`REMOTE<WORKER>: ${remoteWorkerText}`);
+	},
+};