+++ packages/miniflare/CHANGELOG.md
@@ -1,5 +1,36 @@
 # miniflare
 
+## 4.20250525.0
+
+### Minor Changes
+
+- [#9387](https://github.com/cloudflare/workers-sdk/pull/9387) [`e39a45f`](https://github.com/cloudflare/workers-sdk/commit/e39a45ffa0d783cc99107f8ab02d6b3dd27d4c9f) Thanks [@jamesopstad](https://github.com/jamesopstad)! - Add `logReady` method to Miniflare `Log` class. This makes it possible to override the messages printed on server start.
+
+- [#9376](https://github.com/cloudflare/workers-sdk/pull/9376) [`fdae3f7`](https://github.com/cloudflare/workers-sdk/commit/fdae3f7665a5cd3b5e25c9de19156ecd54618a7c) Thanks [@jamesopstad](https://github.com/jamesopstad)! - Add support for Node.js style custom handlers for service bindings and outbound services. This makes it easier to integrate Miniflare with existing Node.js middleware and libraries as `req` and `res` objects can be used directly.
+
+  ```js
+  new Miniflare({
+  	serviceBindings: {
+  		CUSTOM: {
+  			node: (req, res) => {
+  				res.end(`Hello world`);
+  			},
+  		},
+  	},
+  });
+  ```
+
+### Patch Changes
+
+- [#9366](https://github.com/cloudflare/workers-sdk/pull/9366) [`d9d937a`](https://github.com/cloudflare/workers-sdk/commit/d9d937ab6f2868271dde5a8da625773085eaec85) Thanks [@dependabot](https://github.com/apps/dependabot)! - chore: update dependencies of "miniflare" package
+
+  The following dependency versions have been updated:
+
+  | Dependency                | From          | To            |
+  | ------------------------- | ------------- | ------------- |
+  | workerd                   | 1.20250523.0  | 1.20250525.0  |
+  | @cloudflare/workers-types | ^4.20250523.0 | ^4.20250525.0 |
+
 ## 4.20250523.0
 
 ### Minor Changes